bind BalanceHistory service
